+++ src/Server/FileSystemContents/UTF16FileContents.cs
@@ -29,7 +29,7 @@ namespace VsChromium.Server.FileSystemContents {
 
     private IntPtr Pointer { get { return _heap.Pointer; } }
 
-    protected override int CharacterSize {
+    protected override byte CharacterSize {
       get { return sizeof (char); }
     }
